feat(cli): Add cleanup jobs as a cli option
